requirements.txt
numpy
pillow
matplotlib
h5py
streamlit
h5py
scipy